requirements.txt
aiohttp==3.9.0
numpy==1.24.3
faiss-cpu==1.7.4
python-dotenv==1.0.0